refactor: got rid of BillingForm
